make `sqlc.yaml` copy/pastable
Having `sqlc.(yaml|yml)` almost takes up as many characters on the page as it saves, but more importantly, it makes it so that I can't just copy `sqlc.yaml` (or `sqlc.yml`) and get to work. This enumerates the two to simplify reading, search indexing, and copying of the filename.